move config and microservices containers to appropriate folders
